Refactored Blend fixtures
